chore: refactor and condense Route and Measure models to GisResponse
