+++ cv-data-controller/src/main/java/com/trihydro/cvdatacontroller/model/GisResponse/Attributes.java
@@ -1,4 +1,4 @@
-package com.trihydro.cvdatacontroller.model.Route;
+package com.trihydro.cvdatacontroller.model.GisResponse;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -11,4 +11,13 @@ import lombok.Setter;
 public class Attributes {
     @JsonProperty("Route")
     private String Route;
+
+    @JsonProperty("Measure")
+    private Double Measure;
+
+    @JsonProperty("MMin")
+    private Double MMin;
+
+    @JsonProperty("MMax")
+    private Double MMax;
 }
